CMP 141 PROGRAM05
For this assignment, you will create a program that loads the titanic dataset into a pandas DataFrame. The titanic dataset is stored in a CSV file. A description of this dataset will be given in class. Your program will then perform the following tasks:

Task 1: Remove all rows of the dataset where the following features have missing values: age, sex, pclass, embarked, and survived.

Task 2: Create a function called survival_rate that accepts a condition and computes and returns the survival rate for all passengers that satisfy the given condition.

Task 3: Create a list of survival rates based on the following attributes and values: sex=female, sex=male, age<=15, age>15, pclass=1, pclass=2, pclass=3, embarked=C, embarked=Q, embarked=S. Sort this list in descending order.

Task 4: Create a new dataframe with the ‘Condition’ and the corresponding ‘Survival Rate’ as columns.

Task 5: From the dataframe created in the previous task, create a horizontal bar chart of the survival rates with the highest rates on top in descending order (see image below).

Task 6: Add data labels to the bars (see image below).

Task 7: Format the axes labels and the data labels as shown in the image below.